+++ backend/pkg/api/handlers/errorHandler.go
@@ -1,0 +1,7 @@
+package handlers
+
+import "net/http"
+
+func ErrorHandler(w http.ResponseWriter, message string, statusCode int) {
+	http.Error(w, message, statusCode)
+}
