+++ backend/pkg/models/fileTreeitem.go
@@ -1,0 +1,13 @@
+package models
+
+import (
+	"backend/pkg/enums"
+)
+
+type FileTreeItem struct {
+	Id                    string
+	Path                  string
+	Name                  string
+	Type                  enums.FileType
+	AssociatedAudioFileId string
+}
