+++ backend/pkg/utilities/getContentTypeHeaderMimeType.go
@@ -1,0 +1,17 @@
+package utilities
+
+import (
+	"backend/pkg/models"
+	"mime"
+	"path/filepath"
+)
+
+func GetContentTypeHeaderMimeType(file models.FileTreeItem) string {
+	ext := filepath.Ext(file.Path)
+	mimeType := mime.TypeByExtension(ext)
+	if ext == ".vtt" {
+		mimeType = "text/vtt"
+	}
+
+	return mimeType
+}
