--- backend/router/errorHandler.go
@@ -1,7 +1,0 @@
-package router
-
-import "net/http"
-
-func ErrorHandler(w http.ResponseWriter, message string, statusCode int) {
-	http.Error(w, message, statusCode)
-}
