+++ backend/test/README.md
@@ -1,0 +1,1 @@
+TODO INTEGRATION TESTS
